agent/comms/discord_bot/requirements.txt
discord.py>=2.3
httpx>=0.26
structlog>=24.1
pydantic>=2.5
pydantic-settings>=2.1
sqlalchemy>=2.0
asyncpg>=0.29
pgvector>=0.3
redis>=5.0
tiktoken>=0.5
